convert civicrm_groups to a module
