+++ config/initializers/civicrm_verification_workflow.rb
@@ -8,10 +8,3 @@ Decidim::Verifications.register_workflow(:civicrm) do |workflow|
     options.attribute :regional_scope, type: :enum, choices: -> { Decidim::Civicrm::Api::RegionalScope::ALL.keys.map(&:to_s) }
   end
 end
-
-require "decidim/verifications/civicrm_groups"
-
-Decidim::Verifications.register_workflow(:civicrm_groups) do |workflow|
-  workflow.engine = Decidim::Verifications::CivicrmGroups::Engine
-  workflow.admin_engine = Decidim::Verifications::CivicrmGroups::AdminEngine
-end
