+++ decidim-civicrm/Gemfile
@@ -3,24 +3,14 @@
 source "https://rubygems.org"
 
 ruby RUBY_VERSION
-
-DECIDIM_VERSION = { git: "https://github.com/decidim/decidim", branch: "release/0.22-stable" }.freeze
+DECIDIM_VERSION="0.22" # TODO: read from parent
 
 gem "decidim", DECIDIM_VERSION
-# gem "decidim-consultations", DECIDIM_VERSION
-# gem "decidim-initiatives", DECIDIM_VERSION
-gem "decidim-civicrm", path: "./decidim-civicrm"
+gem "decidim-civicrm", path: "."
 
+gem "puma", ">= 4.3"
 gem "bootsnap", "~> 1.4"
-gem "health_check"
-gem "sentry-raven"
-gem "sidekiq", "~> 6.0"
-gem "sidekiq-cron"
-
-gem "puma", "~> 4.3"
 gem "uglifier", "~> 4.1"
-
-gem "faker", "~> 1.9"
 
 group :development, :test do
   gem "byebug", "~> 11.0", platform: :mri
@@ -29,13 +19,10 @@ group :development, :test do
 end
 
 group :development do
+  gem "faker", "~> 1.9"
   gem "letter_opener_web", "~> 1.3"
   gem "listen", "~> 3.1"
   gem "spring", "~> 2.0"
   gem "spring-watcher-listen", "~> 2.0"
   gem "web-console", "~> 3.5"
-end
-
-group :production do
-  gem "fog-aws"
 end
