+++ decidim-civicrm/config/locales/en.yml
@@ -1,0 +1,5 @@
+en:
+  decidim:
+    authorization_handlers:
+      civicrm_groups:
+        name: CiViCRM Groups Authentication
