+++ decidim-civicrm/lib/decidim/civicrm.rb
@@ -1,0 +1,14 @@
+
+# frozen_string_literal: true
+
+require "decidim/civicrm/verifications/civicrm_groups/engine"
+require "decidim/civicrm/verifications/civicrm_groups/admin"
+require "decidim/civicrm/verifications/civicrm_groups/admin_engine"
+require "decidim/civicrm/verifications/civicrm_groups/workflow"
+
+module Decidim
+  # This namespace holds the logic of the `Civicrm` component. This component
+  # allows users to create civicrm in a participatory space.
+  module Civicrm
+  end
+end
