+++ decidim-civicrm/lib/decidim/civicrm/verifications/civicrm_groups/admin.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Civicrm
+    module Verifications
+      module CivicrmGroups
+        # This module contains all the domain logic associated to Decidim's
+        # civicrm group authorization admin panel.
+        module Admin
+        end
+      end
+    end
+  end
+end
