+++ decidim-civicrm/lib/decidim/civicrm/version.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  # This holds the decidim-civicrm version.
+  module Civicrm
+    def self.version
+      "0.22.0"
+    end
+  end
+end
