--- lib/decidim/verifications/civicrm_groups.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require "decidim/verifications/civicrm_groups/admin"
-require "decidim/verifications/civicrm_groups/admin_engine"
-require "decidim/verifications/civicrm_groups/engine"
